Update LoadCell_Scale
Currently this code begins at a near zero value. Once screw begin adding it increases by about 10 per screw added. However this is meant to output in lbs meaning that currently the code is showing a much higher weight value. Additionally if we are attempting to find the current number of screws we need to find the weight of just one screw and then divide the output value by that